+++ Assets/OxGKit/LoggingSystem/Scripts/Runtime/Logging/Enum/ConfigFileType.cs
@@ -1,0 +1,8 @@
+namespace OxGKit.LoggingSystem
+{
+    public enum ConfigFileType
+    {
+        Json,
+        Bytes
+    }
+}